refactor(loaders): move enums into types/enums.py
